requirements.txt
pandas
pyarrow
boto3
python-dotenv